+++ webviz/foxglove/etc/completions/vlink-foxglove.zsh
@@ -45,7 +45,7 @@ _vlink-foxglove() {
         '--proxy_interface=[Proxy interface]:iface:(proxy_api proxy_server)' \
         '--proxy_role=[Proxy role]:role:(controller listener)' \
         '--proxy_domain_id=[Proxy domain id]:id:' \
-        '--proxy_dds_impl=[Proxy DDS impl]:impl:(dds ddsc ddsr ddst)' \
+        '--proxy_dds_impl=[Proxy DDS impl]:impl:(dds ddsc ddsr)' \
         '--proxy_bind_ip=[Proxy bind IP]:ip:' \
         '--proxy_peer_ip=[Proxy peer IP]:ip:' \
         '--proxy_buf_size=[Proxy buffer size]:n:' \